enable different methods to allow graceful shutdown

+++ packages/jobs/src/daemon/index.js
@@ -1,21 +1,52 @@
-import runLoop from './runLoop'
 import range from 'lodash/range'
 import Worker from '../Worker'
 import JobRepository from './JobsRepository'
+import sleep from '../helpers/sleep'
+import loop from './loop'
 import {config} from '@orion-js/app'
 
-export default function ({jobs, workersCount}) {
-  const {logger} = config()
+class Daemon {
+  constructor() {
+    this.running = false
+    this.workers = []
+  }
 
-  JobRepository.setJobs(Object.keys(jobs))
-  JobRepository.deleteUnclaimedJobs()
-    .then(result => {
-      if (result.deletedCount > 0) logger.info(`${result.deletedCount} unclaimed jobs deleted`)
-    })
-    .catch(error => logger.warn('error deleting unclaimed jobs', error))
+  init({jobs, workersCount}) {
+    if (this.running) return
+    this.running = true
+    this.jobs = jobs
+    const {logger} = config()
+    JobRepository.setJobs(jobs)
+    JobRepository.deleteUnclaimedJobs()
+      .then(result => {
+        if (result.deletedCount > 0) logger.info(`${result.deletedCount} unclaimed jobs deleted`)
+      })
+      .catch(error => logger.warn('error deleting unclaimed jobs', error))
 
-  const workers = range(workersCount).map(index => new Worker({index}))
-  runLoop({jobs, workers})
+    this.workers = range(workersCount).map(index => new Worker({index}))
+    global.jobWorkers = this.workers
+    logger.info(`Starting jobs with ${workersCount} workers`)
+    this.run()
+  }
 
-  global.jobWorkers = workers
+  async run() {
+    while (this.running) {
+      this.currentLoop = loop(this)
+      const delay = await this.currentLoop
+      if (delay) {
+        await sleep(delay)
+      }
+    }
+  }
+
+  async stop() {
+    if (!this.running) return
+    this.running = false
+    const {logger} = config()
+    logger.info('Stopping jobs...')
+    await Promise.all([this.currentLoop, ...this.workers.map(worker => worker.close())])
+    logger.info('Jobs stopped')
+  }
 }
+
+export default new Daemon()
